requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.28.0
autogen-agentchat>=0.4.0
autogen-ext
openai>=1.0.0
